migrated test_vtk_reporter_mask
