+++ lettuce/ext/_reporter/vtk_reporter.py
@@ -52,10 +52,10 @@ class VTKReporter(Reporter):
         Usage: vtk_reporter.output_mask(simulation.no_collision_mask)"""
         point_dict = dict()
         if simulation.flow.stencil.d == 2:
-            point_dict["mask"] = simulation.flow.context.convert_to_numpy(
+            point_dict["mask"] = simulation.flow.context.convert_to_ndarray(
                 simulation.no_collision_mask)[..., None].astype(int)
         else:
-            point_dict["mask"] = simulation.flow.context.convert_to_numpy(
+            point_dict["mask"] = simulation.flow.context.convert_to_ndarray(
                 simulation.no_collision_mask).astype(int)
         vtk.gridToVTK(self.filename_base + "_mask",
                       np.arange(0, point_dict["mask"].shape[0]),
